Simplify VectorTileFeature.decodeProperties
Instead of decoding each property to a separate
map, decode to a single map. This reduces Map
allocations, and means that properties can be
resolved with a simple map lookup rather than
iterating through a list of maps and attempting
a lookup in each one.
This improvement reduces memory usage and
makes property lookup more efficient.